# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateTagRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  requests.UpdateTagRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  responses.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive RESTful API built with Go, Fiber, and PostgreSQL using
    Hexagonal Architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go RESTful API - Hexagonal Architecture
  version: "1.0"
paths:
...
  /tags:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all tags with pagination
      parameters:
      - description: 'Number of tags to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of tags to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of tags
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag object
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tag created successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict - Tag with same slug already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - tags
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag by its ID
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag deleted successfully
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Bad request - Invalid tag ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Tag does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - tags
    get:
      consumes:
      - application/json
      description: Retrieve a tag by its ID
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad request - Invalid tag ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Tag does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tag by ID
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Update an existing tag by its ID
      parameters:
      - description: Tag ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tag object
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad request - Invalid tag ID or input data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found - Tag does not exist
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Conflict - Tag with same slug already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - tags
  /tags/search:
    get:
      consumes:
      - application/json
      description: Search tags by query with pagination
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: 'Number of tags to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of tags to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search tags
      tags:
      - tags
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"